+++ BanEvasionAccountDeleteHelper/BanEvasionAccountDeleteHelper.user.ts
@@ -289,8 +289,12 @@
             }
         }
 
-        getDeletionDetails() {
+        validateFields() {
             DeleteEvasionAccountControls.validateLength('Deletion reason details', this.deletionDetails, config.validationBounds.deleteReasonDetails);
+            DeleteEvasionAccountControls.validateLength('Annotation details', this.annotationDetails, config.validationBounds.annotationDetails);
+        }
+
+        getDeletionDetails() {
             return {
                 sockAccountId: this.sockAccountId,
                 deletionDetails: this.deletionDetails
@@ -298,7 +302,6 @@
         }
 
         getAnnotationDetails() {
-            DeleteEvasionAccountControls.validateLength('Annotation details', this.annotationDetails, config.validationBounds.annotationDetails);
             return {
                 mainAccountId: this.mainAccountId,
                 annotationDetails: this.annotationDetails
@@ -312,21 +315,31 @@
         const submitButton = $('<button class="flex--item s-btn s-btn__filled s-btn__danger" type="button">Delete and Annotate</button>');
         submitButton.on('click', (ev) => {
             ev.preventDefault();
-            const {sockAccountId, deletionDetails} = controller.getDeletionDetails();
-            const {mainAccountId, annotationDetails} = controller.getAnnotationDetails();
+            controller.validateFields();
+            void StackExchange.helpers.showConfirmModal({
+                title: 'Are you sure you want to delete this account?',
+                body: 'You will be deleting this account and placing an annotation on the main. This operation cannot be undone.',
+                buttonLabelHtml: 'I\'m sure'
+            })
+                .then(res => {
+                    if (res) {
+                        const {sockAccountId, deletionDetails} = controller.getDeletionDetails();
+                        const {mainAccountId, annotationDetails} = controller.getAnnotationDetails();
 
-            deleteUser(sockAccountId, deletionDetails)
-                .then(() => {
-                    return annotateUser(mainAccountId, annotationDetails);
-                })
-                .then(() => {
-                    // Reload current page if delete and annotation is successful
-                    window.location.reload();
-                    // Open new tab to send message to main account
-                    window.open(`/users/message/create/${mainAccountId}`, '_blank');
-                })
-                .catch(err => {
-                    console.error(err);
+                        deleteUser(sockAccountId, deletionDetails)
+                            .then(() => {
+                                return annotateUser(mainAccountId, annotationDetails);
+                            })
+                            .then(() => {
+                                // Reload current page if delete and annotation is successful
+                                window.location.reload();
+                                // Open new tab to send message to main account
+                                window.open(`/users/message/create/${mainAccountId}`, '_blank');
+                            })
+                            .catch(err => {
+                                console.error(err);
+                            });
+                    }
                 });
         });
         // Build Modal
